requirements.txt
rich
web3
requests
python-dotenv
